refactor(ride-service): refactor code
- switched snake-case to kebab-cases in url path
- removed id in createPromoCode url
- remove /book from ride urls
- switched discount from Double to Integer
- switched cost from Double to BigDecimal
- fixed field names in entities
- removed unnecessary @Mapping's in mappers
- moved JsonSerializing/Deserializing from entity to enums
- added final to utility classes
- removed @BeforeEach and refactor tests
- removed public in test classes

+++ ride-service/src/main/java/com/cabaggregator/rideservice/core/dto/promo/PromoCodeAddingDto.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 
 public record PromoCodeAddingDto(
         String value,
-        Double discount,
+        Integer discount,
         LocalDateTime startDate,
         LocalDateTime endDate
 ) {
