+++ ride-service/src/main/java/com/cabaggregator/rideservice/core/dto/ride/promo/RidePromoCodeDto.java
@@ -2,6 +2,6 @@ package com.cabaggregator.rideservice.core.dto.ride.promo;
 
 public record RidePromoCodeDto(
         String value,
-        Double discount
+        Integer discount
 ) {
 }
